requirements.txt
httpx==0.28.1
livekit==1.1.3
python-dotenv==1.2.2
